+++ Realm/RLMSyncManager.mm
@@ -20,6 +20,7 @@
 
 #import "RLMRealmConfiguration+Sync.h"
 #import "RLMSyncConfiguration_Private.hpp"
+#import "RLMSyncErrorActionToken_Private.hpp"
 #import "RLMSyncSession_Private.hpp"
 #import "RLMSyncUser_Private.hpp"
 #import "RLMSyncUtil_Private.hpp"
@@ -171,29 +172,23 @@ static dispatch_once_t s_onceToken;
     NSError *error = nil;
     BOOL shouldMakeError = YES;
     NSDictionary *custom = nil;
+    // Note that certain types of errors are 'interactive'; users have several options
+    // as to how to proceed after the error is reported.
     switch (errorClass) {
         case RLMSyncSystemErrorKindClientReset: {
-            // Users can respond to "client reset" errors to a
-            // greater degree than possible for most other errors.
-            std::string original_path = [userInfo[@(realm::SyncError::c_original_file_path_key)] UTF8String];
-            custom = @{kRLMSyncPathOfRealmBackupCopyKey: userInfo[@(realm::SyncError::c_recovery_file_path_key)],
-                       kRLMSyncInitiateClientResetBlockKey: ^{
-                           SyncManager::shared().immediately_run_file_actions(original_path);
-                       }};
+            std::string path = [userInfo[@(realm::SyncError::c_original_file_path_key)] UTF8String];
+            custom = @{kRLMSyncPathOfRealmBackupCopyKey:
+                           userInfo[@(realm::SyncError::c_recovery_file_path_key)],
+                       kRLMSyncErrorActionTokenKey:
+                           [[RLMSyncErrorActionToken alloc] initWithOriginalPath:std::move(path)]
+                       };;
             break;
         }
         case RLMSyncSystemErrorKindPermissionDenied: {
-            __block BOOL calledAlready = NO;
-            custom = @{kRLMSyncInitiateDeleteRealmBlockKey: ^ {
-                NSString *originalPath = userInfo[@(realm::SyncError::c_original_file_path_key)];
-                if (calledAlready) {
-                    @throw RLMException(@"The handler block for the Realm at '%@' has already been called once.",
-                                        originalPath);
-                }
-                calledAlready = YES;
-                std::string original_path = [originalPath UTF8String];
-                SyncManager::shared().immediately_run_file_actions(original_path);
-            }};
+            std::string path = [userInfo[@(realm::SyncError::c_original_file_path_key)] UTF8String];
+            custom = @{kRLMSyncErrorActionTokenKey:
+                           [[RLMSyncErrorActionToken alloc] initWithOriginalPath:std::move(path)]
+                       };
             break;
         }
         case RLMSyncSystemErrorKindUser: