+++ RealmSwift/Sync.swift
@@ -82,6 +82,41 @@ public typealias UserCompletionBlock = RLMUserCompletionBlock
  */
 public typealias SyncError = RLMSyncError
 
+extension SyncError {
+    /**
+     An opaque token allowing the user to take action after certain types of
+     errors have been reported.
+
+     - see: `RLMSyncErrorActionToken`
+     */
+    public typealias ActionToken = RLMSyncErrorActionToken
+
+    /**
+     Given a client reset error, extract and return the recovery file path
+     and the action token.
+
+     - see: `SyncError.ActionToken`, `SyncSession.immediatelyHandleError(_:)`
+     */
+    public func clientResetInfo() -> (String, SyncError.ActionToken)? {
+        if code == SyncError.clientResetError,
+            let recoveryPath = userInfo[kRLMSyncPathOfRealmBackupCopyKey] as? String,
+            let token = _nsError.__rlmSync_errorActionToken() {
+            return (recoveryPath, token)
+        }
+        return nil
+    }
+
+    /**
+     Given a permission denied error, extract and return the recovery file path
+     and the action token.
+
+     - see: `SyncError.ActionToken`, `SyncSession.immediatelyHandleError(_:)`
+     */
+    public func deleteRealmUserInfo() -> SyncError.ActionToken? {
+        return _nsError.__rlmSync_errorActionToken()
+    }
+}
+
 /**
  An error associated with network requests made to the authentication server. This type of error
  may be returned in the callback block to `SyncUser.logIn()` upon certain types of failed login
@@ -119,23 +154,6 @@ public typealias SyncManagementObjectStatus = RLMSyncManagementObjectStatus
  - see: `RLMIdentityProvider`
  */
 public typealias Provider = RLMIdentityProvider
-
-public extension SyncError {
-    /// Given a client reset error, extract and return the recovery file path and the reset closure.
-    public func clientResetInfo() -> (String, () -> Void)? {
-        if code == SyncError.clientResetError,
-            let recoveryPath = userInfo[kRLMSyncPathOfRealmBackupCopyKey] as? String,
-            let block = _nsError.__rlmSync_clientResetBlock() {
-            return (recoveryPath, block)
-        }
-        return nil
-    }
-
-    /// Given a permission denied error, extract and return the reset closure.
-    public func deleteRealmUserInfo() -> (() -> Void)? {
-        return _nsError.__rlmSync_deleteRealmBlock()
-    }
-}
 
 /**
  A `SyncConfiguration` represents configuration parameters for Realms intended to sync with